Rename waterfall.js to make_fake_pm_data.js To run in terminal: node make_fake_pm_data.js 'filename.json'
